#3240 Upgraded PHPStan from level 4 to level 5
Fixed all 101 type errors surfaced by the stricter level, covering:
- Auth user type narrowing (Authenticatable|null → User|null) across controllers, middleware, models and Blade views
- floorIndex string/int casts at indexOrFacade() call sites
- JoinClause::on() converted to 3-arg form to accept DB::raw() expressions
- Lua::call() inline ignores for C-extension string-based function names
- with()/load() closure contravariance inline ignores (Larastan limitation)
- DatabaseSeeder SEEDERS constant typed as class-string<TableSeederInterface>
- ChallengeMode __toString() added; getUiMapIds() return type corrected
- Expansion collection cache properties retyped as Collection<string, Season>
- PatreonServiceLogging bool→string for loadAdminUserUpdatedTokenSuccessfully
- CacheServiceLogging Exception→Throwable for PSR cache exception compatibility
- CombatantInfo constructor uses intersection type (CombatantInfoInterface&BaseEvent)
- PollCombatLogRunsCommand buildFilterForCriterion typed to interface; @var narrowing for Dungeon/Collection args
- Various float→int casts, CURLOPT_RETURNTRANSFER 1→true, json_encode true flag removal
- Rule::in() wrapped in array, DungeonRoute::update() replaced with save()
- MDTMappingExportService get('col') → get(['col']), Save.php collection @var assertions

+++ app/Service/CombatLogEvent/Dtos/CombatLogEventFilter.php
@@ -508,8 +508,8 @@ class CombatLogEventFilter implements Arrayable
                 throw new InvalidArgumentException('Mapping version does not have a timer max seconds value');
             }
 
-            $combatLogEventFilter->setDurationMin(($heatmapDataFilter->getTimerFractionMin() * 60) / $timerSeconds);
-            $combatLogEventFilter->setDurationMax(($heatmapDataFilter->getTimerFractionMax() * 60) / $timerSeconds);
+            $combatLogEventFilter->setDurationMin((int)(($heatmapDataFilter->getTimerFractionMin() * 60) / $timerSeconds));
+            $combatLogEventFilter->setDurationMax((int)(($heatmapDataFilter->getTimerFractionMax() * 60) / $timerSeconds));
         }
 
         return $combatLogEventFilter;
